requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astunparse==1.6.3
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
daphne==4.1.2
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.4.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.2
flatbuffers==25.1.24
fonttools==4.56.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
gTTS==2.5.4
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jax==0.5.0
jaxlib==0.5.0
jiter==0.9.0
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.21
ml-dtypes==0.4.1
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
pypdf==5.4.0
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
sentencepiece==0.2.0
service-identity==24.2.0
six==1.17.0
sklearn==0.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sounddevice==0.5.1
sqlparse==0.5.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tomli==2.2.1
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
zope.interface==7.2
